docs: split trees compat server and client modules
