+++ apps/docs/app/trees/tree-examples/SearchSection.tsx
@@ -6,8 +6,9 @@ import { TreeExampleHeading } from '../../components/TreeExampleHeading';
 import { FeatureHeader } from '../../diff-examples/FeatureHeader';
 import { DEFAULT_FILE_TREE_PANEL_CLASS, searchOptions } from './demo-data';
 import { TreeExampleSection } from './TreeExampleSection';
-import { FileTree, preloadFileTree } from '@/lib/treesCompat';
 import { PRODUCTS } from '@/app/product-config';
+import { preloadFileTree } from '@/lib/treesCompat';
+import { FileTree } from '@/lib/treesCompatClient';
 
 const PREPOPULATED_SEARCH = 'tsx';
 const PRESELECTED_FILE = 'src/components/Button.tsx';
