+++ apps/docs/public/trees/llms-full.txt
@@ -119,15 +119,15 @@ folder containing `index.ts`, `utils/helpers.ts`, and a `components` folder with
 | ------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `initialFiles`            | **Required.** Array of file paths (strings). Defines the list and hierarchy.                                                                                                                                                |
 | `id`                      | Unique identifier for this instance (DOM ids, SSR). Default: e.g. `ft_brw_1`.                                                                                                                                               |
-| `flattenEmptyDirectories` | When `true`, single-child folder chains become one row (e.g. `build / assets / images` → `build · assets · images`). Default: `false`. [Live demo](/preview/trees#flattening).                                              |
+| `flattenEmptyDirectories` | When `true`, single-child folder chains become one row (e.g. `build / assets / images` → `build · assets · images`). Default: `false`. [Live demo](../#flattening).                                                         |
 | `fileTreeSearchMode`      | Search behavior: `'expand-matches'` (default), `'collapse-non-matches'`, or `'hide-non-matches'`.                                                                                                                           |
 | `search`                  | Render the built-in search input and enable keyboard-open search behavior. Default: `false`.                                                                                                                                |
 | `useLazyDataLoader`       | When `true`, children load when a folder is expanded. Default: `false`.                                                                                                                                                     |
 | `dragAndDrop`             | Enables built-in drag and drop interactions. Default: `false`.                                                                                                                                                              |
 | `lockedPaths`             | Optional list of file/folder paths that cannot be dragged when drag and drop is enabled.                                                                                                                                    |
 | `onCollision`             | Optional callback for drag collisions. Return `true` to overwrite destination.                                                                                                                                              |
-| `gitStatus`               | Optional `GitStatusEntry[]` used to show Git-style file status (`added`, `modified`, `deleted`). Folders with changed descendants also receive a change indicator. [Live demo](/preview/trees#path-colors).                 |
-| `icons`                   | Optional built-in icon set selection or `FileTreeIconConfig` for semantic colors, CSS-themable palettes, and custom sprite overrides. [Live demo](/preview/trees#custom-icons).                                             |
+| `gitStatus`               | Optional `GitStatusEntry[]` used to show Git-style file status (`added`, `modified`, `deleted`). Folders with changed descendants also receive a change indicator. [Live demo](../#path-colors).                            |
+| `icons`                   | Optional built-in icon set selection or `FileTreeIconConfig` for semantic colors, CSS-themable palettes, and custom sprite overrides. [Live demo](../#custom-icons).                                                        |
 | `sort`                    | Sort children within each directory. `true` (default) uses the standard sort (folders first, dot-prefixed next, case-insensitive alphabetical). `false` preserves insertion order. `{ comparator: fn }` for custom sorting. |
 | `virtualize`              | Enable virtualized rendering so only visible items are rendered. Pass `{ threshold: number }` to activate when item count exceeds the threshold, or `false` to disable. Default: `undefined` (off).                         |
 
@@ -775,8 +775,8 @@ void refreshGitStatus();
 
 Use the `gitStatus` option to show file-level status indicators for added,
 modified, and deleted files. Folders that contain changed descendants
-automatically receive a change hint.
-[Try the live demo](/preview/trees#path-colors) on the homepage.
+automatically receive a change hint. [Try the live demo](../#path-colors) on the
+homepage.
 
 `GitStatusEntry` pairs a file `path` with a status value:
 `{ path: string; status: 'added' | 'modified' | 'deleted' }`.
@@ -800,7 +800,7 @@ target:
 
 Use the `icons` option inside `FileTreeOptions` to choose one of the built-in
 icon sets or inject your own SVG sprite. Try the live demo at
-[/preview/trees#custom-icons](/preview/trees#custom-icons).
+[the homepage](../#custom-icons).
 
 - `icons: 'minimal' | 'standard' | 'complete'` — use one of the shipped icon
   tiers. Each tier is cumulative: `standard` includes everything in `minimal`
